Move farm modules to their own directory.
