+++ src/core/App.vue
@@ -18,7 +18,7 @@ import { refreshCache } from './idb/cache';
 import { loadFieldModules, updateFieldModules } from './field-modules';
 import bundles, { loadConfigDocs, updateConfigDocs } from './store/configDocuments';
 import { alert } from './warnings/alert';
-import profile, { loadProfile, updateProfile } from './store/profile';
+import profile, { loadProfile, updateProfile } from './farm/profile';
 import settings from './shell/settings';
 
 export default {
