+++ src/core/http/interceptor.js
@@ -1,7 +1,7 @@
 import {
   clone, curryN, evolve, partition, reduce,
 } from 'ramda';
-import { getHost } from './remote';
+import { getHost } from '../farm/remote';
 import asArray from '../utils/asArray';
 import Warning from '../warnings/Warning';
